+++ chroma/generator.py
@@ -1,6 +1,7 @@
 from pathlib import Path
 from typing import Callable
 
+from chroma.colors import Color
 from chroma.logger import Logger
 from chroma.utils.dynamic import discover_modules
 from chroma.utils.generator import write_lua_colors
@@ -23,7 +24,17 @@ def prepare():
                 logger.debug(f"Registered generator '{name}'")
 
 
-def generate(name: str, image_path: Path | str, output_path: Path | str, **kwargs):
+def generate(
+    name: str,
+    image_path: Path | str,
+    output_path: Path | str | None = None,
+    **kwargs,
+) -> dict[str, Color] | None:
+    """
+    If `output_path` is None, then the colors are returned as a dict of color
+    name and the resultant color object. Otherwise, the color is converted to
+    a lua theme and written to the output path provided it is valid.
+    """
     prepare()
     generator = GENERATORS_REGISTRY.get(name)
     if generator is None:
@@ -32,5 +43,8 @@ def generate(name: str, image_path: Path | str, output_path: Path | str, **kwarg
     image_path = Path(image_path)
     theme = generator(image_path, **kwargs)
 
-    output_path = Path(output_path)
-    write_lua_colors(output_path, theme)
+    if output_path is None:
+        return theme
+    else:
+        output_path = Path(output_path)
+        write_lua_colors(output_path, theme)