+++ tests/__init__.py
@@ -1,0 +1,1 @@
+# To allow the tests to see the `utils.py` file