+++ tests/fixtures/integrations/gtk.lua
@@ -1,0 +1,53 @@
+theme.gtk = {{
+  colors = {{
+    accent_color = "#abcdef",
+    accent_fg_color = "#abcdef",
+    accent_bg_color = "#abcdef",
+    window_fg_color = "#abcdef",
+    window_bg_color = "#abcdef",
+    view_fg_color = "#abcdef",
+    view_bg_color = "#abcdef",
+    headerbar_fg_color = "#abcdef",
+    headerbar_bg_color = "#abcdef",
+    card_fg_color = "#abcdef",
+    card_bg_color = "#abcdef",
+    dialog_fg_color = "#abcdef",
+    dialog_bg_color = "#abcdef",
+    popover_fg_color = "#abcdef",
+    popover_bg_color = "#abcdef",
+    sidebar_fg_color = "#abcdef",
+    sidebar_bg_color = "#abcdef",
+    backdrop_fg_color = "#abcdef",
+    backdrop_bg_color = "#abcdef",
+  }},
+
+  palettes = {{
+    palette1 = {{
+      blue = "#abcdef",
+      green = "#abcdef",
+      yellow = "#abcdef",
+      orange = "#abcdef",
+      red = "#abcdef",
+      purple = "#abcdef",
+      brown = "#abcdef",
+      light = "#abcdef",
+      dark = "#abcdef",
+    }},
+    palette2 = {{
+      blue = "#abcdef",
+      green = "#abcdef",
+      yellow = "#abcdef",
+      orange = "#abcdef",
+      red = "#abcdef",
+      purple = "#abcdef",
+      brown = "#abcdef",
+      light = "#abcdef",
+      dark = "#abcdef",
+    }},
+  }},
+
+  out = {{
+    gtk3 = "{outpath}/out.gtk3",
+    gtk4 = "{outpath}/out.gtk4",
+  }}
+}}