+++ tests/fixtures/integrations/kitty.lua
@@ -1,0 +1,23 @@
+theme.kitty = {{
+  colors = {{
+    background = "#abcdef",
+    foreground = "#abcdef",
+    black = "#abcdef",
+    red = "#abcdef",
+    green = "#abcdef",
+    yellow = "#abcdef",
+    blue = "#abcdef",
+    magenta = "#abcdef",
+    cyan = "#abcdef",
+    white = "#abcdef",
+    bright_black = "#abcdef",
+    bright_red = "#abcdef",
+    bright_green = "#abcdef",
+    bright_yellow = "#abcdef",
+    bright_blue = "#abcdef",
+    bright_magenta = "#abcdef",
+    bright_cyan = "#abcdef",
+    bright_white = "#abcdef",
+  }},
+  out = "{outpath}/out.kitty"
+}}